fix: fixing and cleaning
