+++ apps/platform-app/src/app/[locale]/(jobs)/(components)/JobDetailsPage/JobDetails.tsx
@@ -10,7 +10,7 @@ import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 import { publishJobAction } from '../../_actions/mutations/publishJob'
 import { type JobModel } from '../../_models/job.model'
-import { storePendingPublishJob } from '../../_utils/jobStorage'
+import { storePendingPublishJob } from '../../_utils/job-storage.client'
 import { AvatarsDisplay } from '../AvatarsDisplay/AvatarsDisplay'
 import { JobDetailsBasicInfo } from '../JobDetailsBasicInfo/JobDetailsBasicInfo'
 import { JobDetailsDetailsInfo } from '../JobDetailsMainInfo.tsx/JobDetailsDetailsInfo'
