+++ apps/platform-app/src/app/[locale]/(jobs)/_utils/job-storage.client.ts
